Studio - Move panel components into src/apps/studio
Relocates the nine remaining StudioWrapper components into src/apps/studio/components. Import paths only; no behaviour change.
