+++ src/apps/studio/components/StudioInspectorPanel.tsx
@@ -17,21 +17,12 @@ import { CloseRounded } from "@mui/icons-material";
 import PhotoLibraryRounded from "@mui/icons-material/PhotoLibraryRounded";
 import AddLinkRounded from "@mui/icons-material/AddLinkRounded";
 import LinkOffRounded from "@mui/icons-material/LinkOffRounded";
-import {
-  ConnectField,
-  ElementSlot,
-} from "../../../../../../../studio/hooks/studioTypes";
-import {
-  NO_TAG,
-  TEXT_TAGS,
-} from "../../../../../../../studio/components/studioTags";
-import { getFieldMeta } from "../../../../../../../studio/components/studioFieldMeta";
-import { FieldIconChip } from "../../../../../../../studio/components/FieldIconChip";
-import {
-  connectFieldToParsley,
-  parseParsleyRef,
-} from "../../../../../../../studio/components/studioParsley";
-import { useCrossModelConnectField } from "../../hooks/useCrossModelConnectField";
+import { ConnectField, ElementSlot } from "../hooks/studioTypes";
+import { NO_TAG, TEXT_TAGS } from "./studioTags";
+import { getFieldMeta } from "./studioFieldMeta";
+import { FieldIconChip } from "./FieldIconChip";
+import { connectFieldToParsley, parseParsleyRef } from "./studioParsley";
+import { useCrossModelConnectField } from "../../content-editor/src/app/views/ItemEdit/hooks/useCrossModelConnectField";
 import { StudioLinkItemDialog } from "./StudioLinkItemDialog";
 
 // Title shown for a supported element tag.
