+++ src/apps/studio/components/StudioPreview.tsx
